pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "literavore"
version = "0.1.0"
description = "Conference paper processing pipeline: fetch, extract, summarize, embed, search"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Lavi Shpigelman"},
]
keywords = ["conference", "papers", "ai", "pipeline", "openreview", "semantic-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "openreview-py>=1.50.0",
    "openai>=1.0.0",
    "pypdf>=4.0.0",
    "pdfplumber>=0.10.0",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "numpy>=1.24.0",
    "mcp[cli]>=1.17.0",
    "streamlit>=1.28.0",
    "pikepdf>=8.0.0",
    "plotly>=5.17.0",
    "umap-learn>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
    "coverage>=7.0",
]

[project.scripts]
literavore = "literavore.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/literavore"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
